+++ docs/history.rst
@@ -3,6 +3,8 @@ History
 
 Latest
 ------
+- DEP: Python 3.7+ (#67)
+- REF: Write grid_mapping to encoding instead of attrs (#66)
 
 0.0.16
 ------